Split Robot class into Model + Robot.
