+++ core/include/jiminy/core/control/AbstractController.h
@@ -27,17 +27,19 @@ namespace jiminy
     //////////////////////////////////////////////////////////////////////////////////////////////
     class AbstractController
     {
-    public:
+    protected:
         ///////////////////////////////////////////////////////////////////////////////////////////////
         /// \brief      Dictionary gathering the configuration options shared between controllers
         ///////////////////////////////////////////////////////////////////////////////////////////////
-        virtual configHolder_t getDefaultOptions()
+        virtual configHolder_t getDefaultControllerOptions()
         {
             configHolder_t config;
             config["telemetryEnable"] = true;
 
             return config;
         };
+
+    public:
         ///////////////////////////////////////////////////////////////////////////////////////////////
         /// \brief      Structure with the configuration options shared between controllers
         ///////////////////////////////////////////////////////////////////////////////////////////////
