+++ core/src/robot/Model.cc
@@ -1,0 +1,736 @@
+
+#include <fstream>
+#include <exception>
+
+#include "pinocchio/parsers/urdf.hpp"
+#include "pinocchio/algorithm/kinematics.hpp"
+#include "pinocchio/algorithm/frames.hpp"
+
+#include "jiminy/core/Utilities.h"
+#include "jiminy/core/Constants.h"
+
+#include "jiminy/core/robot/Model.h"
+
+
+namespace jiminy
+{
+    Model::Model(void) :
+    pncModel_(),
+    pncData_(pncModel_),
+    pncModelRigidOrig_(),
+    pncDataRigidOrig_(pncModelRigidOrig_),
+    mdlOptions_(nullptr),
+    contactForces_(),
+    isInitialized_(false),
+    urdfPath_(),
+    hasFreeflyer_(false),
+    mdlOptionsHolder_(),
+    contactFramesNames_(),
+    contactFramesIdx_(),
+    rigidJointsNames_(),
+    rigidJointsModelIdx_(),
+    rigidJointsPositionIdx_(),
+    rigidJointsVelocityIdx_(),
+    flexibleJointsNames_(),
+    flexibleJointsModelIdx_(),
+    positionLimitMin_(),
+    positionLimitMax_(),
+    velocityLimit_(),
+    positionFieldNames_(),
+    velocityFieldNames_(),
+    accelerationFieldNames_(),
+    pncModelFlexibleOrig_(),
+    nq_(0),
+    nv_(0),
+    nx_(0)
+    {
+        setOptions(getDefaultModelOptions());
+    }
+
+    hresult_t Model::initialize(std::string const & urdfPath,
+                                bool_t      const & hasFreeflyer)
+    {
+        hresult_t returnCode = hresult_t::SUCCESS;
+
+        // Initialize the URDF model
+        returnCode = loadUrdfModel(urdfPath, hasFreeflyer);
+        isInitialized_ = true;
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            // Backup the original model and data
+            pncModelRigidOrig_ = pncModel_;
+            pncDataRigidOrig_ = pinocchio::Data(pncModelRigidOrig_);
+
+            /* Get the list of joint names of the rigid model and
+               remove the 'universe' and 'root' if any, since they
+               are not actual joints. */
+            rigidJointsNames_ = pncModelRigidOrig_.names;
+            rigidJointsNames_.erase(rigidJointsNames_.begin()); // remove the 'universe'
+            if (hasFreeflyer)
+            {
+                rigidJointsNames_.erase(rigidJointsNames_.begin()); // remove the 'root'
+            }
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            // Create the flexible model
+            returnCode = generateModelFlexible();
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            // Add biases to the dynamics properties of the model
+            returnCode = generateModelBiased();
+        }
+
+        if (returnCode != hresult_t::SUCCESS)
+        {
+            // Set the initialization flag
+            isInitialized_ = false;
+        }
+
+        return returnCode;
+    }
+
+    void Model::reset(void)
+    {
+        if (isInitialized_)
+        {
+            // Update the biases added to the dynamics properties of the model.
+            generateModelBiased();
+        }
+    }
+
+    hresult_t Model::addContactPoints(std::vector<std::string> const & frameNames)
+    {
+        if (!isInitialized_)
+        {
+            std::cout << "Error - Model::addContactPoints - Model not initialized." << std::endl;
+            return hresult_t::ERROR_INIT_FAILED;
+        }
+
+        // Make sure that the frame list is not empty
+        if (frameNames.empty())
+        {
+            std::cout << "Error - Model::addContactPoints - The list of frames must not be empty." << std::endl;
+            return hresult_t::ERROR_BAD_INPUT;
+        }
+
+        // Make sure that no frame are duplicates
+        if (checkDuplicates(frameNames))
+        {
+            std::cout << "Error - Model::addContactPoints - Some frames are duplicates." << std::endl;
+            return hresult_t::ERROR_BAD_INPUT;
+        }
+
+        // Make sure that no contact point is associated with any of the frame in the list
+        if (checkIntersection(contactFramesNames_, frameNames))
+        {
+            std::cout << "Error - Model::addContactPoints - At least one of the frame is already been associated with a contact point." << std::endl;
+            return hresult_t::ERROR_BAD_INPUT;
+        }
+
+        // Make sure that all the frames exist
+        for (std::string const & frame : frameNames)
+        {
+            if (!pncModel_.existFrame(frame))
+            {
+                std::cout << "Error - Model::addContactPoints - At least one of the frame does not exist." << std::endl;
+                return hresult_t::ERROR_BAD_INPUT;
+            }
+        }
+
+        // Add the list of frames to the set of contact points
+        contactFramesNames_.insert(contactFramesNames_.end(), frameNames.begin(), frameNames.end());
+
+        // Reset the contact force internal buffer
+        contactForces_ = forceVector_t(contactFramesNames_.size(), pinocchio::Force::Zero());
+
+        // Refresh proxies associated with the contact points only
+        refreshContactsProxies();
+
+        return hresult_t::SUCCESS;
+    }
+
+    hresult_t Model::removeContactPoints(std::vector<std::string> const & frameNames)
+    {
+        if (!isInitialized_)
+        {
+            std::cout << "Error - Model::removeContactPoints - Model not initialized." << std::endl;
+            return hresult_t::ERROR_INIT_FAILED;
+        }
+
+        // Make sure that no frame are duplicates
+        if (checkDuplicates(frameNames))
+        {
+            std::cout << "Error - Model::removeContactPoints - Some frames are duplicates." << std::endl;
+            return hresult_t::ERROR_BAD_INPUT;
+        }
+
+        // Make sure that every frame in the list is associated with a contact point
+        if (!checkInclusion(contactFramesNames_, frameNames))
+        {
+            std::cout << "Error - Model::removeContactPoints - At least one of the frame is not associated with any contact point." << std::endl;
+            return hresult_t::ERROR_BAD_INPUT;
+        }
+
+        // Remove the list of frames from the set of contact points
+        if (!frameNames.empty())
+        {
+            eraseVector(contactFramesNames_, frameNames);
+        }
+        else
+        {
+            contactFramesNames_.clear();
+        }
+
+        // Reset the contact force internal buffer
+        contactForces_ = forceVector_t(contactFramesNames_.size(), pinocchio::Force::Zero());
+
+        // Refresh proxies associated with the contact points only
+        refreshContactsProxies();
+
+        return hresult_t::SUCCESS;
+    }
+
+    hresult_t Model::generateModelFlexible(void)
+    {
+        hresult_t returnCode = hresult_t::SUCCESS;
+
+        if (!isInitialized_)
+        {
+            std::cout << "Error - Model::generateModelFlexible - Model not initialized." << std::endl;
+            returnCode = hresult_t::ERROR_INIT_FAILED;
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            flexibleJointsNames_.clear();
+            flexibleJointsModelIdx_.clear();
+            pncModelFlexibleOrig_ = pncModelRigidOrig_;
+            for(flexibleJointData_t const & flexibleJoint : mdlOptions_->dynamics.flexibilityConfig)
+            {
+                std::string const & jointName = flexibleJoint.jointName;
+
+                // Look if given joint exists in the joint list.
+                if (returnCode == hresult_t::SUCCESS)
+                {
+                    int32_t jointIdx;
+                    returnCode = getJointPositionIdx(pncModel_, jointName, jointIdx);
+                }
+
+                // Add joints to model.
+                if (returnCode == hresult_t::SUCCESS)
+                {
+                    std::string newName =
+                        removeFieldnameSuffix(jointName, "Joint") + FLEXIBLE_JOINT_SUFFIX;
+                    flexibleJointsNames_.emplace_back(newName);
+                    insertFlexibilityInModel(pncModelFlexibleOrig_, jointName, newName); // Ignore return code, as check has already been done.
+                }
+            }
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            getJointsModelIdx(pncModelFlexibleOrig_,
+                              flexibleJointsNames_,
+                              flexibleJointsModelIdx_);
+        }
+
+        return returnCode;
+    }
+
+    hresult_t Model::generateModelBiased(void)
+    {
+        hresult_t returnCode = hresult_t::SUCCESS;
+
+        if (!isInitialized_)
+        {
+            std::cout << "Error - Model::generateModelBiased - Model not initialized." << std::endl;
+            returnCode = hresult_t::ERROR_INIT_FAILED;
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            // Reset the robot either with the original rigid or flexible model
+            if (mdlOptions_->dynamics.enableFlexibleModel)
+            {
+                pncModel_ = pncModelFlexibleOrig_;
+            }
+            else
+            {
+                pncModel_ = pncModelRigidOrig_;
+            }
+
+            for (std::string const & jointName : rigidJointsNames_)
+            {
+                int32_t const jointIdx = pncModel_.getJointId(jointName);
+
+                vector3_t & comRelativePositionBody =
+                    const_cast<vector3_t &>(pncModel_.inertias[jointIdx].lever());
+                comRelativePositionBody +=
+                    randVectorNormal(3U, mdlOptions_->dynamics.centerOfMassPositionBodiesBiasStd);
+
+                // Cannot be less than 1g for numerical stability
+                float64_t & massBody =
+                    const_cast<float64_t &>(pncModel_.inertias[jointIdx].mass());
+                massBody =
+                    std::max(massBody +
+                        randNormal(0.0, mdlOptions_->dynamics.massBodiesBiasStd), 1.0e-3);
+
+                // Cannot be less 1g applied at 1mm of distance from the rotation center
+                vector6_t & inertiaBody =
+                    const_cast<vector6_t &>(pncModel_.inertias[jointIdx].inertia().data());
+                inertiaBody =
+                    clamp(inertiaBody +
+                        randVectorNormal(6U, mdlOptions_->dynamics.inertiaBodiesBiasStd), 1.0e-9);
+
+                vector3_t & relativePositionBody =
+                    pncModel_.jointPlacements[jointIdx].translation();
+                relativePositionBody +=
+                    randVectorNormal(3U, mdlOptions_->dynamics.relativePositionBodiesBiasStd);
+            }
+
+            // Initialize Pinocchio Data internal state
+            pncData_ = pinocchio::Data(pncModel_);
+            pinocchio::forwardKinematics(pncModel_, pncData_,
+                                         vectorN_t::Zero(pncModel_.nq),
+                                         vectorN_t::Zero(pncModel_.nv));
+            pinocchio::updateFramePlacements(pncModel_, pncData_);
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            // Initialize the internal proxies.
+            returnCode = refreshProxies();
+        }
+
+        return returnCode;
+    }
+
+    hresult_t Model::refreshProxies(void)
+    {
+        hresult_t returnCode = hresult_t::SUCCESS;
+
+        if (!isInitialized_)
+        {
+            std::cout << "Error - Model::refreshProxies - Model not initialized." << std::endl;
+            returnCode = hresult_t::ERROR_INIT_FAILED;
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            // Extract the dimensions of the configuration and velocity vectors
+            nq_ = pncModel_.nq;
+            nv_ = pncModel_.nv;
+            nx_ = nq_ + nv_;
+
+            // Extract some rigid joints indices in the model
+            getJointsModelIdx(pncModel_, rigidJointsNames_, rigidJointsModelIdx_);
+            getJointsPositionIdx(pncModel_, rigidJointsNames_, rigidJointsPositionIdx_, false);
+            getJointsVelocityIdx(pncModel_, rigidJointsNames_, rigidJointsVelocityIdx_, false);
+
+            /* Generate the fieldnames associated with the configuration,
+               velocity, and acceleration vectors. */
+            positionFieldNames_.clear();
+            positionFieldNames_.resize(nq_);
+            velocityFieldNames_.clear();
+            velocityFieldNames_.resize(nv_);
+            accelerationFieldNames_.clear();
+            accelerationFieldNames_.resize(nv_);
+            std::vector<std::string> const & jointNames = pncModel_.names;
+            std::vector<std::string> jointShortNames =
+                removeFieldnamesSuffix(jointNames, "Joint");
+            for (uint32_t i=0; i<jointNames.size(); ++i)
+            {
+                std::string const & jointName = jointNames[i];
+                int32_t const jointIdx = pncModel_.getJointId(jointName);
+
+                int32_t idx_q = pncModel_.joints[jointIdx].idx_q();
+
+                if (idx_q >= 0) // Otherwise the joint is not part of the vectorial representation
+                {
+                    int32_t idx_v = pncModel_.joints[jointIdx].idx_v();
+
+                    joint_t jointType;
+                    std::string jointPrefix;
+                    if (returnCode == hresult_t::SUCCESS)
+                    {
+                        returnCode = getJointTypeFromId(pncModel_, jointIdx, jointType);
+                    }
+                    if (returnCode == hresult_t::SUCCESS)
+                    {
+                        if (jointType == joint_t::FREE)
+                        {
+                            // Discard the joint name for FREE joint type since it is unique if any
+                            jointPrefix = FREE_FLYER_PREFIX_BASE_NAME;
+                            jointShortNames[i] = "";
+                        }
+                        else
+                        {
+                            jointPrefix = JOINT_PREFIX_BASE;
+                        }
+                    }
+
+                    std::vector<std::string> jointTypePositionSuffixes;
+                    std::vector<std::string> jointPositionFieldnames;
+                    if (returnCode == hresult_t::SUCCESS)
+                    {
+                        returnCode = getJointTypePositionSuffixes(jointType,
+                                                                  jointTypePositionSuffixes);
+                    }
+                    if (returnCode == hresult_t::SUCCESS)
+                    {
+                        for (std::string const & suffix : jointTypePositionSuffixes)
+                        {
+                            jointPositionFieldnames.emplace_back(
+                                jointPrefix + "Position" + jointShortNames[i] + suffix);
+                        }
+                    }
+                    if (returnCode == hresult_t::SUCCESS)
+                    {
+                        std::copy(jointPositionFieldnames.begin(),
+                                  jointPositionFieldnames.end(),
+                                  positionFieldNames_.begin() + idx_q);
+                    }
+
+                    std::vector<std::string> jointTypeVelocitySuffixes;
+                    std::vector<std::string> jointVelocityFieldnames;
+                    std::vector<std::string> jointAccelerationFieldnames;
+                    if (returnCode == hresult_t::SUCCESS)
+                    {
+                        returnCode = getJointTypeVelocitySuffixes(jointType,
+                                                                  jointTypeVelocitySuffixes);
+                    }
+                    if (returnCode == hresult_t::SUCCESS)
+                    {
+                        for (std::string const & suffix : jointTypeVelocitySuffixes)
+                        {
+                            jointVelocityFieldnames.emplace_back(
+                                jointPrefix + "Velocity" + jointShortNames[i] + suffix);
+                            jointAccelerationFieldnames.emplace_back(
+                                jointPrefix + "Acceleration" + jointShortNames[i] + suffix);
+                        }
+                    }
+                    if (returnCode == hresult_t::SUCCESS)
+                    {
+                        std::copy(jointVelocityFieldnames.begin(),
+                                  jointVelocityFieldnames.end(),
+                                  velocityFieldNames_.begin() + idx_v);
+                        std::copy(jointAccelerationFieldnames.begin(),
+                                  jointAccelerationFieldnames.end(),
+                                  accelerationFieldNames_.begin() + idx_v);
+                    }
+                }
+            }
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            // Get the joint position limits from the URDF or the user options
+            if (mdlOptions_->joints.positionLimitFromUrdf)
+            {
+                positionLimitMin_.resize(rigidJointsPositionIdx_.size());
+                positionLimitMax_.resize(rigidJointsPositionIdx_.size());
+                for (uint32_t i=0; i < rigidJointsPositionIdx_.size(); ++i)
+                {
+                    positionLimitMin_[i] = pncModel_.lowerPositionLimit[rigidJointsPositionIdx_[i]];
+                    positionLimitMax_[i] = pncModel_.upperPositionLimit[rigidJointsPositionIdx_[i]];
+                }
+            }
+            else
+            {
+                positionLimitMin_ = mdlOptions_->joints.positionLimitMin;
+                positionLimitMax_ = mdlOptions_->joints.positionLimitMax;
+            }
+
+            // Get the joint velocity limits from the URDF or the user options
+            if (mdlOptions_->joints.velocityLimitFromUrdf)
+            {
+                velocityLimit_.resize(rigidJointsVelocityIdx_.size());
+                for (uint32_t i=0; i < rigidJointsVelocityIdx_.size(); ++i)
+                {
+                    velocityLimit_[i] = pncModel_.velocityLimit[rigidJointsVelocityIdx_[i]];
+                }
+            }
+            else
+            {
+                velocityLimit_ = mdlOptions_->joints.velocityLimit;
+            }
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            returnCode = refreshContactsProxies();
+        }
+
+        return returnCode;
+    }
+
+    hresult_t Model::refreshContactsProxies(void)
+    {
+        hresult_t returnCode = hresult_t::SUCCESS;
+
+        if (!isInitialized_)
+        {
+            std::cout << "Error - Model::refreshContactsProxies - Model not initialized." << std::endl;
+            returnCode = hresult_t::ERROR_INIT_FAILED;
+        }
+
+        if (returnCode == hresult_t::SUCCESS)
+        {
+            // Extract the contact frames indices in the model
+            getFramesIdx(pncModel_, contactFramesNames_, contactFramesIdx_);
+        }
+
+        return returnCode;
+    }
+
+    hresult_t Model::setOptions(configHolder_t modelOptions)
+    {
+        bool_t internalBuffersMustBeUpdated = false;
+        bool_t isFlexibleModelInvalid = false;
+        bool_t isCurrentModelInvalid = false;
+        if (isInitialized_)
+        {
+            /* Check that the following user parameters has the right dimension,
+               then update the required internal buffers to reflect changes, if any. */
+            configHolder_t & jointOptionsHolder =
+                boost::get<configHolder_t>(modelOptions.at("joints"));
+            if (!boost::get<bool_t>(jointOptionsHolder.at("positionLimitFromUrdf")))
+            {
+                vectorN_t & positionLimitMin = boost::get<vectorN_t>(jointOptionsHolder.at("positionLimitMin"));
+                if ((int32_t) rigidJointsPositionIdx_.size() != positionLimitMin.size())
+                {
+                    std::cout << "Error - Model::setOptions - Wrong vector size for 'positionLimitMin'." << std::endl;
+                    return hresult_t::ERROR_BAD_INPUT;
+                }
+                vectorN_t positionLimitMinDiff = positionLimitMin - mdlOptions_->joints.positionLimitMin;
+                internalBuffersMustBeUpdated |= (positionLimitMinDiff.array().abs() >= EPS).all();
+                vectorN_t & positionLimitMax = boost::get<vectorN_t>(jointOptionsHolder.at("positionLimitMax"));
+                if ((uint32_t) rigidJointsPositionIdx_.size() != positionLimitMax.size())
+                {
+                    std::cout << "Error - Model::setOptions - Wrong vector size for 'positionLimitMax'." << std::endl;
+                    return hresult_t::ERROR_BAD_INPUT;
+                }
+                vectorN_t positionLimitMaxDiff = positionLimitMax - mdlOptions_->joints.positionLimitMax;
+                internalBuffersMustBeUpdated |= (positionLimitMaxDiff.array().abs() >= EPS).all();
+            }
+            if (!boost::get<bool_t>(jointOptionsHolder.at("velocityLimitFromUrdf")))
+            {
+                vectorN_t & velocityLimit = boost::get<vectorN_t>(jointOptionsHolder.at("velocityLimit"));
+                if ((int32_t) rigidJointsVelocityIdx_.size() != velocityLimit.size())
+                {
+                    std::cout << "Error - Model::setOptions - Wrong vector size for 'velocityLimit'." << std::endl;
+                    return hresult_t::ERROR_BAD_INPUT;
+                }
+                vectorN_t velocityLimitDiff = velocityLimit - mdlOptions_->joints.velocityLimit;
+                internalBuffersMustBeUpdated |= (velocityLimitDiff.array().abs() >= EPS).all();
+            }
+
+            // Check if the flexible model and its associated proxies must be regenerated
+            configHolder_t & dynOptionsHolder =
+                boost::get<configHolder_t>(modelOptions.at("dynamics"));
+            bool_t const & enableFlexibleModel = boost::get<bool_t>(dynOptionsHolder.at("enableFlexibleModel"));
+            flexibilityConfig_t const & flexibilityConfig =
+                boost::get<flexibilityConfig_t>(dynOptionsHolder.at("flexibilityConfig"));
+
+            if (mdlOptions_
+            && (flexibilityConfig.size() != mdlOptions_->dynamics.flexibilityConfig.size()
+                || !std::equal(flexibilityConfig.begin(),
+                                flexibilityConfig.end(),
+                                mdlOptions_->dynamics.flexibilityConfig.begin())))
+            {
+                isFlexibleModelInvalid = true;
+            }
+            else if (mdlOptions_ && enableFlexibleModel != mdlOptions_->dynamics.enableFlexibleModel)
+            {
+                isCurrentModelInvalid = true;
+            }
+        }
+
+        // Update the internal options
+        mdlOptionsHolder_ = modelOptions;
+
+        // Create a fast struct accessor
+        mdlOptions_ = std::make_unique<modelOptions_t const>(mdlOptionsHolder_);
+
+        if (isFlexibleModelInvalid)
+        {
+            // Force flexible model regeneration
+            generateModelFlexible();
+        }
+
+        if (isFlexibleModelInvalid || isCurrentModelInvalid)
+        {
+            // Trigger biased model regeneration
+            reset();
+        }
+        else if (internalBuffersMustBeUpdated)
+        {
+            // Update the info extracted from the model
+            refreshProxies();
+        }
+
+        return hresult_t::SUCCESS;
+    }
+
+    configHolder_t Model::getOptions(void) const
+    {
+        return mdlOptionsHolder_;
+    }
+
+    bool_t const & Model::getIsInitialized(void) const
+    {
+        return isInitialized_;
+    }
+
+    std::string const & Model::getUrdfPath(void) const
+    {
+        return urdfPath_;
+    }
+
+    bool_t const & Model::getHasFreeflyer(void) const
+    {
+        return hasFreeflyer_;
+    }
+
+    hresult_t Model::loadUrdfModel(std::string const & urdfPath,
+                                   bool_t      const & hasFreeflyer)
+    {
+        if (!std::ifstream(urdfPath.c_str()).good())
+        {
+            std::cout << "Error - Model::loadUrdfModel - The URDF file does not exist. Impossible to load it." << std::endl;
+            return hresult_t::ERROR_BAD_INPUT;
+        }
+
+        urdfPath_ = urdfPath;
+        hasFreeflyer_ = hasFreeflyer;
+
+        try
+        {
+            pinocchio::Model pncModel;
+            if (hasFreeflyer)
+            {
+                pinocchio::urdf::buildModel(urdfPath,
+                                            pinocchio::JointModelFreeFlyer(),
+                                            pncModel);
+            }
+            else
+            {
+                pinocchio::urdf::buildModel(urdfPath, pncModel);
+            }
+            pncModel_ = pncModel;
+        }
+        catch (std::exception& e)
+        {
+            std::cout << "Error - Model::loadUrdfModel - Something is wrong with the URDF. Impossible to build a model from it." << std::endl;
+            return hresult_t::ERROR_BAD_INPUT;
+        }
+
+        return hresult_t::SUCCESS;
+    }
+
+    std::vector<std::string> const & Model::getContactFramesNames(void) const
+    {
+        return contactFramesNames_;
+    }
+
+    std::vector<int32_t> const & Model::getContactFramesIdx(void) const
+    {
+        return contactFramesIdx_;
+    }
+
+    std::vector<std::string> const & Model::getPositionFieldNames(void) const
+    {
+        return positionFieldNames_;
+    }
+
+    vectorN_t const & Model::getPositionLimitMin(void) const
+    {
+        return positionLimitMin_;
+    }
+
+    vectorN_t const & Model::getPositionLimitMax(void) const
+    {
+        return positionLimitMax_;
+    }
+
+    std::vector<std::string> const & Model::getVelocityFieldNames(void) const
+    {
+        return velocityFieldNames_;
+    }
+
+    vectorN_t const & Model::getVelocityLimit(void) const
+    {
+        return velocityLimit_;
+    }
+
+    std::vector<std::string> const & Model::getAccelerationFieldNames(void) const
+    {
+        return accelerationFieldNames_;
+    }
+
+    std::vector<std::string> const & Model::getRigidJointsNames(void) const
+    {
+        return rigidJointsNames_;
+    }
+
+    std::vector<int32_t> const & Model::getRigidJointsModelIdx(void) const
+    {
+        return rigidJointsModelIdx_;
+    }
+
+    std::vector<int32_t> const & Model::getRigidJointsPositionIdx(void) const
+    {
+        return rigidJointsPositionIdx_;
+    }
+
+    std::vector<int32_t> const & Model::getRigidJointsVelocityIdx(void) const
+    {
+        return rigidJointsVelocityIdx_;
+    }
+
+    std::vector<std::string> const & Model::getFlexibleJointsNames(void) const
+    {
+        static std::vector<std::string> const flexibleJointsNamesEmpty {};
+        if (mdlOptions_->dynamics.enableFlexibleModel)
+        {
+            return flexibleJointsNames_;
+        }
+        else
+        {
+            return flexibleJointsNamesEmpty;
+        }
+    }
+
+    std::vector<int32_t> const & Model::getFlexibleJointsModelIdx(void) const
+    {
+        static std::vector<int32_t> const flexibleJointsModelIdxEmpty {};
+        if (mdlOptions_->dynamics.enableFlexibleModel)
+        {
+            return flexibleJointsModelIdx_;
+        }
+        else
+        {
+            return flexibleJointsModelIdxEmpty;
+        }
+    }
+
+    uint32_t const & Model::nq(void) const
+    {
+        return nq_;
+    }
+
+    uint32_t const & Model::nv(void) const
+    {
+        return nv_;
+    }
+
+    uint32_t const & Model::nx(void) const
+    {
+        return nx_;
+    }
+}
