+++ examples/double_pendulum/double_pendulum.cc
@@ -66,10 +66,10 @@ int main(int argc, char_t * argv[])
     std::vector<std::string> motorJointNames{"SecondPendulumJoint"};
 
     auto robot = std::make_shared<Robot>();
-    configHolder_t mdlOptions = robot->getOptions();
-    boost::get<bool_t>(boost::get<configHolder_t>(mdlOptions.at("joints")).at("positionLimitFromUrdf")) = true;
-    boost::get<bool_t>(boost::get<configHolder_t>(mdlOptions.at("joints")).at("velocityLimitFromUrdf")) = true;
-    robot->setOptions(mdlOptions);
+    configHolder_t modelOptions = robot->getModelOptions();
+    boost::get<bool_t>(boost::get<configHolder_t>(modelOptions.at("joints")).at("positionLimitFromUrdf")) = true;
+    boost::get<bool_t>(boost::get<configHolder_t>(modelOptions.at("joints")).at("velocityLimitFromUrdf")) = true;
+    robot->setModelOptions(modelOptions);
     robot->initialize(urdfPath, false);
     for (std::string const & jointName : motorJointNames)
     {
@@ -85,8 +85,8 @@ int main(int argc, char_t * argv[])
     controller->initialize(robot);
 
     // Instantiate and configuration the engine
-    Engine engine;
-    configHolder_t simuOptions = engine.getDefaultOptions();
+    auto engine = std::make_shared<Engine>();
+    configHolder_t simuOptions = engine->getOptions();
     boost::get<bool_t>(boost::get<configHolder_t>(simuOptions.at("telemetry")).at("enableConfiguration")) = true;
     boost::get<bool_t>(boost::get<configHolder_t>(simuOptions.at("telemetry")).at("enableVelocity")) = true;
     boost::get<bool_t>(boost::get<configHolder_t>(simuOptions.at("telemetry")).at("enableAcceleration")) = true;
@@ -108,8 +108,8 @@ int main(int argc, char_t * argv[])
     boost::get<float64_t>(boost::get<configHolder_t>(simuOptions.at("contacts")).at("frictionDry")) = 5.0;
     boost::get<float64_t>(boost::get<configHolder_t>(simuOptions.at("contacts")).at("frictionViscous")) = 5.0;
     boost::get<float64_t>(boost::get<configHolder_t>(simuOptions.at("contacts")).at("transitionEps")) = 0.001;
-    engine.setOptions(simuOptions);
-    engine.initialize(robot, controller, callback);
+    engine->setOptions(simuOptions);
+    engine->initialize(robot, controller, callback);
 
     timer.toc();
 
@@ -124,17 +124,17 @@ int main(int argc, char_t * argv[])
 
     // Run simulation
     timer.tic();
-    engine.simulate(tf, x0);
+    engine->simulate(tf, x0);
     timer.toc();
     std::cout << "Simulation time: " << (timer.dt * 1.0e3) << "ms" << std::endl;
 
     // Write the log file
     std::vector<std::string> header;
     matrixN_t log;
-    engine.getLogData(header, log);
+    engine->getLogData(header, log);
     std::cout << log.rows() << " log points" << std::endl;
-    engine.writeLogTxt(outputDirPath + std::string("log.txt"));
-    engine.writeLogBinary(outputDirPath + std::string("log.data"));
+    engine->writeLogTxt(outputDirPath + std::string("log.txt"));
+    engine->writeLogBinary(outputDirPath + std::string("log.data"));
 
     return 0;
 }
